[NUMBER-AS-LONG]  Added better handling of numeric columns
Problem?

When exporting from an oracle database that uses a `number` column type for
bigints, opencsv put `.0` as a suffix to all the values, even though the
values in the database were integer values.

Solution!

Added checking of the scale on numeric datatypes so that if we
encountered a `NUMERIC` jdbc datatype / oracle `number` field, with 0 scale,
we treat the value as `long` instead of `double`.